consolidate info popover and style them closer to designs
