+++ tailwind.config.js
@@ -175,6 +175,7 @@ module.exports = {
         "560px": "560px",
       },
       colors: {
+        "tooltip-bg": "#FDF5DB",
         "maroon-base": "#3C0C0C",
         "kusama-base": "#E10178",
         "kusama-green": "#00E7BD",
